+++ IcVibracoes.Core/NumericalIntegrationMethods/Newmark/INewmarkMethod.cs
@@ -1,28 +1,7 @@
-﻿using IcVibracoes.Core.DTO;
-using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElement;
-using System.Threading.Tasks;
-
-namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark
+﻿namespace IcVibracoes.Core.NumericalIntegrationMethods.Newmark
 {
     /// <summary>
     /// It's responsible to execute the Newmark numerical integration method to calculate the vibration.
     /// </summary>
-    public interface INewmarkMethod
-    {
-        /// <summary>
-        /// Calculates the result for the initial time.
-        /// </summary>
-        /// <param name="input"></param>
-        /// <returns></returns>
-        Task<FiniteElementResult> CalculateResultForInitialTime(FiniteElementMethodInput input);
-
-        /// <summary>
-        /// Calculates and write in a file the response matrixes.
-        /// </summary>
-        /// <param name="input"></param>
-        /// <param name="previousResult"></param>
-        /// <param name="time"></param>
-        /// <returns></returns>
-        Task<FiniteElementResult> CalculateResult(FiniteElementMethodInput input, FiniteElementResult previousResult, double time);
-    }
+    public interface INewmarkMethod : INumericalIntegrationMethod { }
 }