+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/BeamWithDva/CalculateBeamWithDvaVibration.cs
@@ -12,7 +12,6 @@ using IcVibracoes.Core.Mapper;
 using IcVibracoes.Core.Models;
 using IcVibracoes.Core.Models.BeamCharacteristics;
 using IcVibracoes.Core.Models.Beams;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark;
 using IcVibracoes.DataContracts.FiniteElement.BeamWithDynamicVibrationAbsorber;
 using System;
 using System.IO;
@@ -43,7 +42,6 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
         /// <param name="mainMatrix"></param>
         /// <param name="file"></param>
         /// <param name="time"></param>
-        /// <param name="newmarkMethod"></param>
         /// <param name="naturalFrequency"></param>
         public CalculateBeamWithDvaVibration(
             IBoundaryCondition boundaryCondition,
@@ -53,9 +51,8 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
             IBeamWithDvaMainMatrix<TProfile> mainMatrix,
             IFile file,
             ITime time,
-            INewmarkMethod newmarkMethod,
             INaturalFrequency naturalFrequency)
-            : base(file, time, newmarkMethod, naturalFrequency)
+            : base(file, time, naturalFrequency)
         {
             this._boundaryCondition = boundaryCondition;
             this._arrayOperation = arrayOperation;