+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/BeamWithDva/Rectangular/CalculateRectangularBeamWithDvaVibration.cs
@@ -7,7 +7,6 @@ using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithDva.Rectangular;
 using IcVibracoes.Core.Calculator.NaturalFrequency;
 using IcVibracoes.Core.Calculator.Time;
 using IcVibracoes.Core.Mapper;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark;
 
 namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithDva.Rectangular
 {
@@ -26,19 +25,16 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
         /// <param name="mainMatrix"></param>
         /// <param name="file"></param>
         /// <param name="time"></param>
-        /// <param name="newmarkMethod"></param>
         /// <param name="naturalFrequency"></param>
         public CalculateRectangularBeamWithDvaVibration(
-            IBoundaryCondition boundaryCondition, 
+            IBoundaryCondition boundaryCondition,
             IArrayOperation arrayOperation, IRectangularGeometricProperty geometricProperty,
             IMappingResolver mappingResolver,
-            IRectangularBeamWithDvaMainMatrix mainMatrix, 
-            IFile file, 
-            ITime time, 
-            INewmarkMethod newmarkMethod, 
-            INaturalFrequency naturalFrequency) 
-            : base(boundaryCondition, arrayOperation, geometricProperty, mappingResolver, mainMatrix, file, time, newmarkMethod, naturalFrequency)
-        {
-        }
+            IRectangularBeamWithDvaMainMatrix mainMatrix,
+            IFile file,
+            ITime time,
+            INaturalFrequency naturalFrequency)
+            : base(boundaryCondition, arrayOperation, geometricProperty, mappingResolver, mainMatrix, file, time, naturalFrequency)
+        { }
     }
 }