+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/BeamWithPiezoelectric/CalculateBeamWithPiezoelectricVibration.cs
@@ -11,7 +11,6 @@ using IcVibracoes.Core.Mapper;
 using IcVibracoes.Core.Models;
 using IcVibracoes.Core.Models.BeamCharacteristics;
 using IcVibracoes.Core.Models.Beams;
-using IcVibracoes.Core.NumericalIntegrationMethods.Newmark;
 using IcVibracoes.DataContracts.FiniteElement.BeamWithPiezoelectric;
 using System;
 using System.IO;
@@ -43,7 +42,6 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithP
         /// <param name="mainMatrix"></param>
         /// <param name="file"></param>
         /// <param name="time"></param>
-        /// <param name="newmarkMethod"></param>
         /// <param name="naturalFrequency"></param>
         public CalculateBeamWithPiezoelectricVibration(
             IBoundaryCondition boundaryCondition,
@@ -53,9 +51,8 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithP
             IBeamWithPiezoelectricMainMatrix<TProfile> mainMatrix,
             IFile file, 
             ITime time, 
-            INewmarkMethod newmarkMethod, 
             INaturalFrequency naturalFrequency) 
-            : base(file, time, newmarkMethod, naturalFrequency)
+            : base(file, time, naturalFrequency)
         {
             this._boundaryCondition = boundaryCondition;
             this._arrayOperation = arrayOperation;