+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Matrix with piezoelectric boundary conditions to consider the beam fastening.
 - Constructor of class FiniteElementsMethodInput receives the Numerical Method choosen for the analysis to set the correct values on properties Beta and Gama.
 - All names with FiniteElements to FiniteElement.
+- Numerical method instantiated on process operation instead of in the constructor.
 ### Removed
 - All classes RequestData.
 - Classes CentralDifferenceMethodInput, ImplicitAccelerationMethodInput and NewmarkMethodInput.